+++ src/main/java/gift/member/business/service/MemberService.java
@@ -22,15 +22,10 @@ public class MemberService {
 
     private final MemberRepository memberRepository;
     private final JwtGenerator jwtGenerator;
-    private final WishlistRepository wishlistRepository;
-    private final ProductRepository productRepository;
 
-    public MemberService(MemberRepository memberRepository, JwtGenerator jwtGenerator,
-        WishlistRepository wishlistRepository, ProductRepository productRepository) {
+    public MemberService(MemberRepository memberRepository, JwtGenerator jwtGenerator) {
         this.memberRepository = memberRepository;
         this.jwtGenerator = jwtGenerator;
-        this.wishlistRepository = wishlistRepository;
-        this.productRepository = productRepository;
     }
 
     public JwtToken registerMember(MemberRegisterDto memberRegisterDto) {
@@ -45,40 +40,5 @@ public class MemberService {
             throw new LoginException(ErrorCode.LOGIN_ERROR, "패스워드가 이메일과 일치하지 않습니다.");
         }
         return jwtGenerator.createToken(member.getId());
-    }
-
-    public List<WishListDto> getWishLists(Long memberId) {
-        var wishLists = wishlistRepository.getWishListByMemberId(memberId);
-
-        var productIds = wishLists.stream()
-            .map(Wishlist::getProductId)
-            .toList();
-
-        Map<Long, Product> products = productRepository.getProductsByIds(productIds);
-
-        return wishLists.stream()
-            .map(wishlist -> WishListDto.of(
-                wishlist.getId(),
-                products.get(wishlist.getProductId()),
-                wishlist.getCount()
-            ))
-            .toList();
-    }
-
-    public Long addWishList(Long memberId, Long productId) {
-        var wishList = new Wishlist(productId, memberId, 1);
-        return wishlistRepository.saveWishList(wishList);
-    }
-
-    public Long updateWishList(Long memberId, WishlistUpdateDto wishListUpdateDto) {
-        var wishList = wishlistRepository.getWishListByMemberIdAndProductId(
-            memberId, wishListUpdateDto.productId());
-
-        wishList.setCount(wishListUpdateDto.count());
-        return wishlistRepository.updateWishlist(wishList);
-    }
-
-    public void deleteWishList(Long memberId, Long productId) {
-        wishlistRepository.deleteWishlist(memberId, productId);
     }
 }
